hospital changes

+++ GraphColoring/GraphColoringCPU/Algorithm.h
@@ -13,10 +13,10 @@ namespace version_cpu
 	__declspec(dllexport) unsigned int GetFirstBitPosition(int);
 	__declspec(dllexport) unsigned int GetBitPosition(int, int);
 	__declspec(dllexport) int** CreateVertices(int, int);
-	__declspec(dllexport) int* BuildingIndependentSets_BitVersion(int*,int*, int);
-	__declspec(dllexport) int* BuildingIndependentSets_TableVersion(int*,int*, int*, int);
+	__declspec(dllexport) int* BuildingIndependentSets_BitVersion(int*, int*, int);
+	__declspec(dllexport) int* BuildingIndependentSets_TableVersion(int*, int*, int*, int);
 
-	extern "C" __declspec(dllexport) int FindChromaticNumber(int*,int*, int*, int, int=0);
+	extern "C" __declspec(dllexport) int FindChromaticNumber(double*, int*, int*, int, int = 0);
 }
 
 #endif
